remove style cop warnings, move resolver to Framework.Cqrs project
