+++ samples/Framework.Cqrs.Samples/Framework.Cqrs.Samples.csproj
@@ -13,4 +13,8 @@
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.0.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="AspNetCoreResolver\" />
+  </ItemGroup>
+
 </Project>
