+++ src/Framework.Cqrs/Commands/ICommandHandlerWithResponse.cs
@@ -1,6 +1,5 @@
 namespace PetProjects.Framework.Cqrs.Commands
 {
-
     public interface ICommandHandlerWithResponse<in TCommand, out TResponse>
         where TCommand : ICommand
     {
